Map system update
Created a function that automatically disables the map when the object the script is attached to is enabled. Seems to work for both dialogue choices AND regular dialogue. Theoretically works for ANY object. it's called scr_DialogueDisable. You will need to manually point it at the btn_map object I made, and any other buttons you want to disable.